requirements.txt
Django==1.10.3
djangorestframework==3.5.4
celery==4.0.2
redis==2.10.5
psycopg2==2.6.2
django_compressor==2.1.1
requests==2.13.0
